external/webkit/LayoutTests/platform/android/webgl-conformance/dumpAsText/conformance/glsl/implicit/function_int_float.vert-expected.txt

CONSOLE MESSAGE: line 24: *** Error compiling shader '[object WebGLShader]':ERROR: 0:7: 'foo' : no matching overloaded function found 

implicit cast of int to float in function argument should fail

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

test: implicit cast of int to float in function argument should fail
test vertex shader
default fragment shaderPASS implicit cast of int to float in function argument should fail
PASS successfullyParsed is true

TEST COMPLETE
